[BE #26] 장바구니, 회원 등 레포지토리 테스트 및 삭제 메서드 수정

See merge request s10-bigdata-dist-sub2/S10P22A606!17

+++ back/mallook/src/main/java/io/ssafy/mallook/domain/order/entity/ProductHistory.java
@@ -25,6 +25,6 @@ public class ProductHistory extends BaseEntity {
     private String productImage;
     private String productSize;
     private String productColor;
-    private Long productFee;
+    private Integer productFee;
 //    private Shoppingmall shoppingmallId;
 }
